fixed the delete button. Everything is working now.
