+++ src/javascripts/components/backToBoards/backToBoards.js
@@ -1,0 +1,7 @@
+import printB from '../printBoards/printBoards';
+
+const backToUsersBoards = () => {
+  printB.makeABoard();
+};
+
+export default { backToUsersBoards };
