--- src/javascripts/components/buildTheSite/buildTheSite.scss
@@ -1,4 +1,0 @@
-.header {
-  margin: 20px;
-  padding: 20px;
-}
