+++ src/javascripts/components/printBoards/printBoards.scss
@@ -1,0 +1,8 @@
+.header {
+  margin: 20px;
+  padding: 20px;
+}
+
+.singleBoard {
+  *{pointer-events: none;}
+}
